so-middleware/store-app/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
pyarrow==14.0.2
boto3==1.34.34
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-kafka-python==0.42b0
opentelemetry-instrumentation-boto3sqs==0.42b0
fastparquet==2023.10.1
wrapt==1.14.1